requirements.txt
# LangChain & LangGraph (RAG + agent workflow)
langchain>=0.3.0
langchain-anthropic>=0.2.0
langchain-community>=0.3.0
langgraph>=0.2.0

# Embeddings & vector store (RAG retrieval)
chromadb>=0.5.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
sentence-transformers>=3.0.0

# Anthropic (Claude for prompt enhancement)
anthropic>=0.39.0

# Optional: OpenAI embeddings or other OpenAI APIs
langchain-openai>=0.2.0
openai>=1.0.0

# Image generation in exercises uses NVIDIA API + requests (see exercises/image_gen.py)
requests>=2.28.0

# Environment and utilities
python-dotenv>=1.0.0

# Optional: for a simple web UI later
# streamlit>=1.28.0
